feat: consolidate logic into single externally callable action
In order to make this a more "javascript-centric" action many of
the steps that were handled in a workflow have been moved into the
JS code of this action. This makes the implementation of this more
of a standard js action and does not have the anti-pattern of the
workflow calling itself.

This commit also adds documentation such that the action can be used
by others and adds semantic release.

Closes #7

Breaking-change: action is no longer called by remote workflow